remove `--server.endpoint` from startup scripts
Previously, _some_ startup shell scripts that are used instead of the
actual executables contained an explicit mentioning of
`--server.endpoint` with the value `tcp://127.0.0.1:8529`.
This is the default endpoint of ArangoDB, and it works well if the
default endpoint is not changed.
When the server-side endpoint is changed to something else, one can
invoke the tools with `--server.endpoint <custom-value>`, which is fine.
However, what doesn't work is to specify a custom endpoint in a custom
config file. This is because the value `--server.endpoint
tcp://127.0.0.1:8529` will always take precedence over what is in custom
config files. The reason is that command-line arguments are prioritized
and the underlying options are "sealed" once they are processed.
Overwriting them with custom values from a config file then does not
work anymore.
The solution is to not specify a value for `--server.endpoint` in these
shell scripts. The tools will still work, because the default endpoint
is baked into all executables and will be used if no endpoint was
specified.
For some tools, e.g. arangorestore and arangobench, we seem to have
removed the endpoint startup option from the command-lines already. But
we didn't do it consistently for all client-tools. This PR now does
that.